test: connecting to influxdb
